+++ src/main/java/ch/hslu/cas/msed/blobfish/board/EvalBar.java
@@ -9,7 +9,7 @@ class EvalBar {
      * A positive number (e.g., +1.5) means White has an advantage equal to about one and a half pawns;
      * a negative number (e.g., -2.0) means Black has the edge.
      */
-    static int getEvaluation(String positionFen) {
+    static double getEvaluation(String positionFen) {
         throw new UnsupportedOperationException("Not supported yet.");
     }
 }
